server/pyproject.toml
[tool.poetry]
name = "googly-eyes-server"
version = "0.1.0"
description = ""
authors = ["Alex Haslam <alexhaslam12@gmail.com>"]

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = "~3.11"
numpy = "1.23.5"
tensorflow = "^2.15.0"
pillow = "9.5.0"
flask = "^3.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
